requirements.txt
streamlit
pymongo
python-dotenv
streamlit-calendar